feat: add encounter creation with all sections of SCT form
