+++ api/src/main/java/org/openmrs/module/sespct/api/PedidoService.java
@@ -1,12 +1,13 @@
 package org.openmrs.module.sespct.api;
 
-import org.openmrs.Patient;
-import org.openmrs.api.OpenmrsService;
-import org.openmrs.module.sespct.api.model.Pedido;
-
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+
+import org.openmrs.Patient;
+import org.openmrs.api.OpenmrsService;
+import org.openmrs.module.sespct.api.model.Pedido;
+import org.openmrs.module.sespct.api.model.Resposta;
 
 public interface PedidoService extends OpenmrsService {
 	
@@ -25,6 +26,10 @@ public interface PedidoService extends OpenmrsService {
 	void deletePedido(Pedido pedido);
 	
 	Patient mapIdentifier(String patientUuid, Pedido pedido);
+	
+	List<Resposta> getRespostasPendentes();
+	
+	Resposta saveResposta(Resposta resposta);
 	
 	void createDummyData();
