+++ api/src/main/java/org/openmrs/module/sespct/api/dao/impl/RespostaDaoImpl.java
@@ -2,28 +2,35 @@ package org.openmrs.module.sespct.api.dao.impl;
 
 import java.util.List;
 
-import org.hibernate.SessionFactory;
+import org.hibernate.Query;
+import org.openmrs.api.db.hibernate.DbSession;
+import org.openmrs.api.db.hibernate.DbSessionFactory;
 import org.openmrs.module.sespct.api.dao.RespostaDao;
 import org.openmrs.module.sespct.api.model.Resposta;
 
 public class RespostaDaoImpl implements RespostaDao {
 	
-	private SessionFactory sessionFactory;
+	private DbSessionFactory dbSessionFactory;
 	
-	public void setSessionFactory(SessionFactory sessionFactory) {
-		this.sessionFactory = sessionFactory;
+	public void setDbSessionFactory(DbSessionFactory dbSessionFactory) {
+		this.dbSessionFactory = dbSessionFactory;
+	}
+	
+	private DbSession getCurrentSession() {
+		return dbSessionFactory.getCurrentSession();
 	}
 	
 	@Override
 	@SuppressWarnings("unchecked")
 	public List<Resposta> getRespostasPendentes() {
-		String hql = "from Resposta r where r.sincronizado = false";
-		return sessionFactory.getCurrentSession().createQuery(hql).list();
+		final String hql = "from Resposta r where r.sincronizado = false";
+		Query query = getCurrentSession().createQuery(hql);
+		return query.list();
 	}
 	
 	@Override
 	public Resposta saveResposta(Resposta resposta) {
-		sessionFactory.getCurrentSession().saveOrUpdate(resposta);
+		getCurrentSession().saveOrUpdate(resposta);
 		return resposta;
 	}
 }
