+++ api/src/main/resources/liquibase.xml
@@ -779,8 +779,18 @@
 	        ALTER TABLE sespct_linha_solicitada 
 	        MODIFY COLUMN linha VARCHAR(50);
 	    </sql>
-</changeSet>
+	</changeSet>
 	
+	<changeSet id="sespct-1.0.0-15" author="sespct">
+	    <preConditions onFail="MARK_RAN">
+	        <columnExists tableName="sespct_pedido" columnName="pedidoId"/>
+	    </preConditions>
+	    <comment>Renomear coluna pedidoId para NCFT na tabela sespct_pedido</comment>
 	
-    
+	    <renameColumn 
+	        tableName="sespct_pedido"
+	        oldColumnName="pedidoId"
+	        newColumnName="ncft"
+	        columnDataType="varchar(50)"/>
+	</changeSet>
 </databaseChangeLog>
