+++ api/src/main/resources/moduleApplicationContext.xml
@@ -6,7 +6,7 @@
                            http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
                            http://www.springframework.org/schema/context
                            http://www.springframework.org/schema/context/spring-context-3.0.xsd">
-
+                           
     <!-- Enable annotation-based configuration -->
     <context:component-scan base-package="org.openmrs.module.sespct" />
 
@@ -15,24 +15,21 @@
         <property name="dbSessionFactory" ref="dbSessionFactory" />
     </bean>
     
-    <bean id="sespct.RespostaDao" class="org.openmrs.module.sespct.api.dao.impl.RespostaDaoImpl">
-    	<property name="sessionFactory" ref="sessionFactory"/>
-	</bean>
+    <bean id="respostaDao" class="org.openmrs.module.sespct.api.dao.impl.RespostaDaoImpl"> 
+    	<property name="dbSessionFactory" ref="dbSessionFactory" />
+    </bean>
 
     <!-- Define the service implementation -->
-    <bean id="pedidoServiceImpl"
-          class="org.openmrs.module.sespct.api.impl.PedidoServiceImpl">
+    <bean id="pedidoServiceImpl" class="org.openmrs.module.sespct.api.impl.PedidoServiceImpl">
         <property name="pedidoDao" ref="pedidoDao" />
+        <property name="respostaDao" ref="respostaDao" />
     </bean>
-    
-    <bean id="sespct.RespostaService" class="org.openmrs.module.sespct.api.impl.RespostaServiceImpl">
-    	<property name="respostaDao" ref="sespct.RespostaDao"/>
-	</bean>
 
     <!-- Register the service with OpenMRS Context -->
-    <bean parent="serviceContext">
+ 	<bean parent="serviceContext">
         <property name="moduleService">
             <list>
+                <!-- PedidoService Registration -->
                 <value>org.openmrs.module.sespct.api.PedidoService</value>
                 <ref bean="pedidoServiceImpl" />
             </list>
